Updated 7/06/2024

+++ E-EstateV2 API/E-EstateV2 API/DTO/DTO_Field.cs
@@ -4,7 +4,7 @@
     {
         public int id { get; set; }
         public string fieldName { get; set; }
-        public int area { get; set; }
+        public float area { get; set; }
         public bool isMature { get; set; }
         public bool isActive { get; set; }
         public DateTime? dateOpenTapping { get; set; }
